Signed-off-by: Nitishkumar Singh <nitishkumarsingh71@gmail.com>
Add support to run e2e test on docker swarm and kubernetes

we will be able to run the test on both docker and k8, before
any new changes are being merged to the project. This way we
will be able to ensure that all changes are compatible with
both enviornment.